Fix #57: Add support for the top level tags object.
Create folder objects for tags and prefer those tag folders for
methods. Fall back to the path folders when no tags are present
in the OAS3 structure.

Add a final step to prune empty folders from the collection. Supporting
top level tags resulted in a number of empty folders in practice. This
last pass prunes those empty folders. This results in a nice collection
when tags are present in the OAS3 input.