Add the alwaysInheritAuthentication option
This option forces all requests to inherit auth from the top of the
collection. This is really useful when all requests should be using the
same authentication. If authentication is called out on each request in
the Postman app, they will not inherit the options set on the
collection, like the token, and must all be set manually. By allowing
all requests to simply inherit from the collection, the token can be set
once at the top of the collection and all requests will use it
automatically.

Fixes #271